[수정] 거리 조회 시 정수 단위

+++ src/main/java/com/nnzz/nnzz/dto/StoreDTO.java
@@ -15,7 +15,7 @@ public class StoreDTO {
     private String storeId; // 가게 id
     private String name; // 가게 이름
     private String address; // 가게 주소
-    private double distance; // 사용자와의 거리
+    private Integer distance; // 사용자와의 거리
     private String category; // 냠냠쩝쩝 카테고리
     private List<BroadcastDTO> broadcasts; // 방송 정보
     private List<MenuDTO> menus; // 메뉴 정보
